update argument implementation
