+++ demisto_sdk/commands/content_graph/parsers/base_script.py
@@ -1,5 +1,4 @@
 import re
-from dataclasses import dataclass
 from functools import cached_property
 from pathlib import Path
 from typing import List, Optional, Set
@@ -16,14 +15,6 @@ from demisto_sdk.commands.prepare_content.integration_script_unifier import (
 EXECUTE_CMD_PATTERN = re.compile(
     r"execute_?command\(['\"]([a-zA-Z-_]+)['\"].*", re.IGNORECASE
 )
-
-
-@dataclass
-class Argument:
-    name: str
-    description: str
-    deprecated: bool = False
-    default_value: Optional[str] = None
 
 
 class BaseScriptParser(IntegrationScriptParser, content_type=ContentType.BASE_SCRIPT):
@@ -64,18 +55,6 @@ class BaseScriptParser(IntegrationScriptParser, content_type=ContentType.BASE_SC
 
         for cmd in self.get_command_executions():
             self.add_command_or_script_dependency(cmd)
-
-    @property
-    def arguments(self) -> List[Argument]:
-        return [
-            Argument(
-                name=argument.get("name", ""),
-                deprecated=argument.get("deprecated", False),
-                description=argument.get("description"),
-                default_value=argument.get("defaultValue"),
-            )
-            for argument in self.yml_data.get("args", [])
-        ]
 
     @property
     def code(self) -> Optional[str]:
